requirements.txt
torch
torchvision
torch-directml
opencv-python==4.10.0.84
streamlit==1.31.0
streamlit-webrtc==0.47.1
aiortc==1.9.0
plotly==5.18.0
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
av==12.0.0
